Mark mermaid.min.js as binary so it doesn't show up in grep
mermaid.min.js contains a giant blob of minified JavaScript. If a
`git grep` happens to match inside it, it'll display the entire file as
a result, disrupting the ability to grep for anything else. Use
`.gitattributes` to mark it as binary, so it just shows up as "Binary
file mermaid.min.js matches" instead.