s/champion/liaison/
In the new process proposal, I switched from "liaison" to "champion",
but I'm somewhat regretting that. The liaison term has some continuity
from the initiatives idea and it I think captures the more limited
nature of the role -- it's clear that a liaison is not the one
"doing the work", in some sense, but rather rep'ing the lang team POV.

I like champion because TC39 and because I do see the champion as
the one who are psyched about it and who are advocating for it within
the lang team.