Switch to calendar ICS file hosted on GitHub pages
Some people like having a calendar invite for our meetings.  We've
been using Google Calendar to achieve this.  However, this calendar
can be troublesome to update and has often lagged our real schedule.

Other teams were having this same issue, so @davidtwco graciously
created a pair of projects to solve this problem:

- https://github.com/rust-lang/calendar
- https://github.com/rust-lang/calendar-generation

The tooling here allows for our calendar to be specified in TOML and
tracked in git.  From this, it generates an `ical` ICS file that is
hosted on GitHub pages.

All of the lang team and subteam meetings that were previously on the
other calendar have been added to this new calendar.  Let's switch
over to it.

(People will, unfortunately, have to make a one-time update to their
calendars with this new URL.)

To make our meetings more accessible to people who don't view the
calendar, let's go ahead and describe the time of our recurring slots
directly on the calendar page.  We'll describe these in terms of
U.S. time zones because we track U.S. daylight saving time.

Let's also update this page to note that design meeting minutes can be
found via the issue for each design meeting, and let's link to our
documents on HackMD that include the more recent minutes until they
are archived to git.

Finally, we had linked on this page to the style team's calendar.
However, the links to that calendar now return HTTP error 404.  Let's
remove those apparently dead links.